found errors in substr

+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: stitovsk <stitovsk@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/12 00:41:48 by stitovsk          #+#    #+#             */
-/*   Updated: 2023/09/21 16:17:56 by stitovsk         ###   ########.fr       */
+/*   Updated: 2023/09/25 21:26:24 by stitovsk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int count;
+	size_t count;
 
 	count = 0;
 	while (str[count] != '\0')
